Change coordinate data types.
- Change to using u32s instead of unsigned shorts so as to take
  advantage of the native word size of the GBA CPU.

+++ source/line.c
@@ -5,7 +5,7 @@
 int main () {
     REG_DISPCNT = DCNT_MODE3 | DCNT_BG2;
     
-    unsigned short x, y;
+    u32 x, y;
     for (y = 0; y < SCREEN_HEIGHT; y++) {
         for(x = 0; x < SCREEN_WIDTH; x++){
             m3_plot(x, y, x*x + y*y);
